backend/requirements.txt
# Core
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
psycopg[binary]==3.2.3

# Async / Queue
celery[redis]==5.4.0
django-celery-beat==2.7.0
redis==5.2.1

# File & storage
django-storages[s3]==1.14.4
boto3==1.35.92
Pillow==11.0.0

# Documents
openpyxl==3.1.5
xlsxwriter==3.2.0
weasyprint==63.1

# Utils
python-dotenv==1.0.1
gunicorn==23.0.0

# Dev / test
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
freezegun==1.5.1
ruff==0.8.4
mypy==1.13.0
django-stubs==5.1.1
djangorestframework-stubs==3.15.1
ipython==8.30.0
